app/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-admin-sortable2==2.1.3
django-bootstrap4==22.2
django-cleanup==6.0.0
django-environ==0.9.0
easy-thumbnails==2.8.3
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.5
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
